added:CI/CD

+++ tests/integration/mongoose-memory.test.ts
@@ -14,7 +14,9 @@ describe("mongodb-memory-server integration", () => {
   })
 
   afterEach(async () => {
-    await mongoose.connection.db.dropDatabase()
+    if (mongoose.connection.db) {
+      await mongoose.connection.db.dropDatabase()
+    }
   })
 
   afterAll(async () => {
